Tests/LibWeb/Text/expected/css/style-attribute-no-op-invalidation-counters.txt

setting identical style text: 1 invalidations
setting changed style text: 1 invalidations
setting identical custom property: 0 invalidations
setting changed custom property: 1 invalidations
